initial build. working on connecting to db
