--- fns-hl7-pipeline/fn-cosmos-worker/buildNumber.properties
@@ -1,3 +1,0 @@
-#maven.buildNumber.plugin properties file
-#Tue Aug 01 15:39:00 CDT 2023
-buildNumber=16
