+++ spikes/fn-cosmos-worker/.gitignore
