+++ spikes/fn-cosmos-worker/host.json
