+++ spikes/fn-cosmos-worker/src/main/resources/logback.xml
