+++ spikes/fn-cosmos-worker/src/test/kotlin/FunctionTest.kt
@@ -1,0 +1,11 @@
+import org.junit.jupiter.api.Test
+import com.google.gson.GsonBuilder
+import com.google.gson.JsonObject
+import com.google.gson.JsonParser
+
+class FunctionTest {
+    @Test
+    fun cosmosPull(){
+
+    }
+}
